+++ app/Modules/Attachments/Domain/Exceptions/AttachmentNotAvailableException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Modules\Attachments\Domain\Exceptions;
-
-use RuntimeException;
-
-final class AttachmentNotAvailableException extends RuntimeException {}